requirements.txt
# Core dependencies for Ingredient Health Analyzer
streamlit==1.31.0
google-cloud-vision==3.7.0
langchain==0.1.9
langchain-google-vertexai==0.1.0
langchain-community==0.0.24

# Additional dependencies
python-dotenv==1.0.1
tavily-python==0.3.0
